% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractionl.R
\name{Ops.fractional}
\alias{Ops.fractional}
\title{Method for the group generic function for the arithmetic operators}
\usage{
\method{Ops}{fractional}(e1, e2)
}
\arguments{
\item{e1}{A numeric objet, possibly of class \code{"fractional"}}

\item{e2}{A numeric objet, possibly of class \code{"fractional"}}
}
\value{
The result of the arithmetic operation, flagged as class
\code{"fractional"}
}
\description{
Provides arithmetic operations for numeric objects or of class \code{"fractional"}.
}
\examples{
(M <- fractional(1:10/7))
M + 1
1 + M + M^2

}

